Enum trillium_tokio::Binding

pub enum Binding<T, U> {
    Tcp(T),
    Unix(U),
}

A wrapper enum that has blanket implementations for common traits like TryFrom, Stream, AsyncRead, and AsyncWrite. This can contain listeners (like TcpListener), Streams (like Incoming), or bytestreams (like TcpStream).

Variants

Tcp(T)

a tcp type (listener or incoming or stream)

Unix(U)

a unix type (listener or incoming or stream)

Implementations

impl<T, U> Binding<T, U>

pub fn get_tcp(&self) -> Option<&T>

borrows the tcp stream or listener, if this is a tcp variant

pub fn get_unix(&self) -> Option<&U>

borrows the unix stream or listener, if this is unix variant

pub fn get_tcp_mut(&mut self) -> Option<&mut T>

mutably borrows the tcp stream or listener, if this is tcp variant

pub fn get_unix_mut(&mut self) -> Option<&mut U>

mutably borrows the unix stream or listener, if this is unix variant