Function trillium_tokio::run_async

pub async fn run_async(handler: impl Handler)

Runs a trillium handler in an async context with default config

Run the provided trillium handler on an already-running tokio runtime with default settings. The defaults are the same as crate::run. To customize these settings, see crate::config.

This function will poll pending until the server shuts down.